changes in the database design

+++ controllers/comment.js
@@ -1,4 +1,3 @@
-const User = require('../models/users')
 const Comment = require('../models/comments')
 const Post = require('../models/posts')
 const Like = require('../models/likes')
@@ -10,27 +9,14 @@ const { ObjectId } = require('mongodb')
 
 const addComment = async (req, res, next) => {
     try {
-        // const myMedia = ['Post', 'Comment']
         const { describtion } = req.body
         const { id } = req.params
         if (!ObjectId.isValid(id)) throw new AppError('invalid id', 400)
-        
+
         const post = await Post.findById(id)
         if (!post) throw new AppError('post not found', 404)
-        // function fileHandling(arr) {
-        //     if (arr.length > 1) return next(new AppError('only one file can be uploaded', 400))
-        //     else if (arr.length == 1) {
-        //         const file = arr[0]
-        //         const fileName = file.originalname + '_' + Date.now()
-        //         const destinationPath = path.join(__dirname, '../uploads/posts', fileName)
-        //         fs.writeFileSync(destinationPath, file.buffer)
-        //         return fileName
-        //     }
-        //     else if (!describtion) return next(new AppError('comment data missing', 400))
-        // }
-        // const fileName = fileHandling(req.files)
-        // const arr = [id]
         await Comment.create({ 'file': req.file.filename, 'user_id': req.user._id, describtion, 'post_id': id })
+        await Post.findByIdAndUpdate(id, { $inc: { comments: 1 } })
 
         res.json('Comment created successfully')
     }
@@ -56,97 +42,28 @@ const updateComment = async (req, res, next) => {
 
 const getComments = async (req, res, next) => {
     try {
-        
+
         const skip = parseInt(req.query.skip)
         if (!skip && skip != 0) throw new AppError('skip value needed', 400)
         const limit = 3
         const { id } = req.params
         if (!ObjectId.isValid(id)) throw new AppError('invalid post id', 400)
-       
+
         const objectId = new ObjectId(id);
-        const comments = await Comment.aggregate([
-            {
-                $match: { 'post_id': objectId }
-            },
-            {
-                $lookup: {
-                    from: "likes",
-                    let: { comment_id: "$_id", user_id: req.user?._id },
-                    pipeline: [
-                        {
-                            $match: {
-                                $expr: {
-                                    $and: [
-                                        { $eq: ["$comment_id", "$$comment_id"] },
-                                        { $eq: ["$user_id", "$$user_id"] }
-                                    ]
-                                }
-                            }
-                        }
-                    ],
-                    as: "liked"
-                }
-            },
-            {
-                $lookup: {
-                    from: 'likes', // The name of the Likes collection
-                    localField: '_id', // Field from the Posts collection
-                    foreignField: 'comment_id', // Field from the Likes collection
-                    as: 'likes' // Output array field name
-                }
-            },
-            {
-                $lookup: {
-                    from: 'users', // The name of the Likes collection
-                    localField: 'user_id', // Field from the Posts collection
-                    foreignField: '_id', // Field from the Likes collection
-                    as: 'user' // Output array field name
-                }
-            },
-            { $unwind: '$user' },
-
-            
-            {
-                $addFields: {
-                    totalLikes: { $size: '$likes' },  username: '$user.name', username: '$user.name', liked: {
-                        $cond: {
-                            if: { $eq: [{ $size: '$liked' }, 1] },
-                            then: true,
-                            else: false 
-                        }
-                    }  }
-            },
-           
-            {
-                $project: {
-                    likes: 0, // Exclude the 'likes' array from the output
-                    user: 0,
-                    user_id: 0,
-                    post_id: 0
-                }
-            },
-
-            {
-                $sort: { createdAt: -1 }
-            },
-            { $skip: skip }, // Skip the first 10 documents
-
-            // Limit the number of documents returned
-            { $limit: limit } // 
-        ])
-        // await Comment.populate(comments, [{ path: 'users' }])
-
-        // Iterate over each post
+        const comments = await Comment.find({ post_id: id }, { post_id: false }).populate({
+            path: 'user_id',
+            select: 'name' // Include only `name` and `email` fields from `User`
+        })
         comments.forEach(comment => {
-            if (comment.file) comment.file = `${req.protocol}://${req.get('host')}/commentfile/${comment.file}`;
+            comment.file = `${req.protocol}://${req.get('host')}/commentfile/${comment.file}`;
         });
-        // const hah = await Comment.find().populate('user_id')
-        res.send(comments);
+        res.json(comments)
+
     }
     catch (err) { next(err) }
 }
 
-const deleteComment = async (req,res,next) => {
+const deleteComment = async (req, res, next) => {
     try {
         const { id } = req.params;
 
@@ -169,26 +86,31 @@ const deleteComment = async (req,res,next) => {
             throw new AppError('Unauthorized', 401);
         }
 
-        // Delete all likes associated with the post
-        await Like.deleteMany({ comment_id: id });
-
-        const replies = await Reply.find({ comment_id: id });
-
+        // Delete all likes associated with the comment
+        await Like.deleteMany({ comment_id: comment._id });
+        // get all replies related to the comment
+        const replies = await Reply.find({ comment_id: comment._id });
+        // loop through the replies to delete it and its related documents
         for (const reply of replies) {
             // Delete all likes associated with the reply
             await Like.deleteMany({ reply_id: reply._id });
 
             // Delete the reply
             await Reply.findByIdAndDelete(reply._id);
-            fs.unlinkSync(path.join(__dirname, '../uploads/posts', reply.file))
+            // delete the reply file if exists
+            if (reply.file) fs.unlinkSync(path.join(__dirname, '../uploads/replies', reply.file))
+            // decrease the count of total comments of the post
+            await Post.findByIdAndUpdate(comment.post_id, { $inc: { comments: -1 } })
         }
-
-        // Finally, delete the post itself
-        await Comment.findByIdAndDelete(id);
-        fs.unlinkSync(path.join(__dirname, '../uploads/comments', comment.file))
-        res.json('Comment and associated data successfully deleted');
-}
-catch(err) {next(err)}
+        // Delete the comment document
+        await Comment.findByIdAndDelete(comment._id);
+        // delete the comment file
+        if (comment.file) fs.unlinkSync(path.join(__dirname, '../uploads/comments', comment.file))
+        // decrease the count of total comments of the post
+        await Post.findByIdAndUpdate(comment.post_id, { $inc: { comments: -1 } })
+        res.json('Comment successfully deleted');
+    }
+    catch (err) { next(err) }
 }
 
 module.exports = { addComment, getComments, deleteComment, updateComment }
